Update README.md
# Personal Portfolio

This is a personal portfolio showcasing my projects, skills, and experience. It provides an overview of my work and background, highlighting my expertise and interests. Feel free to explore the different sections to learn more about me and my work.

## Table of Contents

- [About Me](#about-me)
- [Projects](#projects)
- [Skills](#skills)
- [Experience](#experience)
- [Contact](#contact)

## About Me

I am a passionate and dedicated individual with a keen interest in technology and innovation. I love solving problems and creating solutions that make a difference. My background in [your field/industry] has equipped me with valuable skills and insights that I bring to every project I work on.

## Projects

Here are some of the projects I have worked on:

1. **Project Name**: Brief description of the project.
   - Technologies Used: List of technologies used.
   - [Link to Project](#) (if applicable)

2. **Project Name**: Brief description of the project.
   - Technologies Used: List of technologies used.
   - [Link to Project](#) (if applicable)

## Skills

- Programming Languages: List of languages.
- Tools & Technologies: List of tools and technologies.
- Soft Skills: List of soft skills.

## Experience

- **Company Name**: Brief description of your role and responsibilities.
  - Duration: Start date - End date

- **Company Name**: Brief description of your role and responsibilities.
  - Duration: Start date - End date

## Contact

- Email: charugoel204@gmail.com
- LinkedIn: https://www.linkedin.com/in/charu-goel-25b646202/
- GitHub: https://github.com/Charu761

Feel free to reach out to me if you have any questions or would like to connect. Thank you for visiting my portfolio!